feat: ex04 finish

+++ TP01/src/ex04.h
@@ -1,7 +1,7 @@
 #ifndef TP01_EX04_H
 #define TP01_EX04_H
 
-int getAverage(int grades[], int n);
+float getAverage(int grades[], int n);
 int getMin(int grades[], int n);
 int getMax(int grades[], int n);
